+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * 2.5.9
     - fixed issue for istanbul source and source path replacement
+    - fixed issue for original file when fully uncovered
 
 * 2.5.8
     - added new option `all` for adding empty coverage for all files