+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+* 2.5.10
+    - fixed coverage for invalid original code
+    - fixed coverage for empty bytes
+    - added `getSnapshot` API for snapshot testing
+
 * 2.5.9
     - fixed issue for istanbul source and source path replacement
     - fixed issue for original file when fully uncovered